refactor(fast_corners): de-duplicate fast_corners module (#47)
Third de-duplication step of #47, following the #62/#63 pattern.

- src/fast_corners/fast_corners.ts: replace the type-only stub with the
  REAL implementation moved verbatim from the monolith (set_threshold,
  detect, _cmp_offsets; imports _cmp_score_16 from ./fast_private, which
  was already a real module).
- src/jsfeatNext.ts: shrinks by ~225 lines; attaches fast_corners from its
  module.

Verified behavior-preserving: tsc --noEmit clean; npm test 57/57 (detector
parity suite pins fast_corners.detect against original jsfeat); UMD bundle
smoke-checked (instanceof chain, static inheritance, corner detection on a
synthetic image).
